+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [TD-4463] Generate audit events for `rule_created` and
   `implementation_created`
+- [TD-4425] Shift Implementations CSV last execution date field timezone
 
 ### Fixed